Reduce size of generated HTML and allow YAML to override library
Libraries are now listed by license. This greatly reduces the size of
the generated HTML-file.
YAML can now override the deduced library. This is to add support for
proprietary licenses with open-source attribution and to fix broken
license report from the repository serving the library.
Added a proprietary template so that proprietary libraries with open-
source attribution can be shown.